requirements.txt
factory-boy==2.12.0
flask==1.0.3
flask-sqlalchemy==2.4.0
psycopg2==2.8.2
sqlalchemy==1.3.3
